+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- The `untag` will remove tags from previous events
+
 ## [0.2.1] - 2019-08-31
 ### Added
 - `summary --refs` will list an Event's reference in the tags